book_data->BookData book_storer->BookStorer NaverBookSearcher is distributed from book_data crawler is now a class
BookData is Facade of my software now
